+++ mod.ts
@@ -1,4 +1,5 @@
 export type { Predicate } from "./sources/predicate.ts";
 export { Repository } from "./sources/repository.ts";
-export { UUIDRepository } from "./sources/specifics/uuid_repository.ts";
+export { StrategyRepository } from "./sources/strategies/strategy_repository.ts";
+export { UUIDRepository } from "./sources/strategies/uuid_repository.ts";
 export type { Updater } from "./sources/updater.ts";